[Skip ci] Multiple updates due to travel

+++ scripts/pda_management.ps1
@@ -1,9 +1,10 @@
 function pda{
     param(
-        [parameter(ValueFromPipelineByPropertyName, HelpMessage = "Please, inform the notification type")][ValidateSet("nk","hl")][string]$Client,
+        [parameter(ValueFromPipelineByPropertyName, HelpMessage = "Please, inform the notification type")][ValidateSet("nk","hl","centerlar")][string]$Client,
         [parameter(ValueFromPipelineByPropertyName, HelpMessage = "Please, inform the notification type")][ValidateSet("Admin","WMS","INV","Etiquetas","Label","etq","apk","Mobile")][string]$Project
         )
-    if($Client -eq "nk"){
+	$env:PDA_ENVIRONMENT = "prod";
+    if($($Client -eq "nk") -and $($env:PDA_ENVIRONMENT -eq "dev")){
         if($Project -eq "Admin" -or $Project -eq "admin" -or $Project -eq "ADMIN"){
             Stop-IISSite -Name NK.Front.Admin
             Expand-Archive -Path $env:USERPROFILE\Desktop\admin.zip -DestinationPath C:\inetpub\wwwroot\repos\NKStore\Cloud.Autentication.React -Force
@@ -17,7 +18,7 @@ function pda{
             Start-IISSite -Name NK.Front.WMS
         }
     }
-    if($Client -eq "hl"){
+    if($($Client -eq "hl") -and $($env:PDA_ENVIRONMENT -eq "dev")){
         if($Project -eq "Admin" -or $Project -eq "admin" -or $Project -eq "ADMIN"){
             Stop-IISSite -Name NK.Front.Admin
             Expand-Archive -Path $env:USERPROFILE\Desktop\admin.zip -DestinationPath C:\inetpub\wwwroot\repos\NKStore\Cloud.Autentication.React -Force
@@ -29,7 +30,39 @@ function pda{
 			Remove-Item -Path $env:USERPROFILE\Desktop\etiquetas.zip -Force
         }
 		if($Project -eq "mobile" -or $Project -eq "apk"){
-            copy-item $env:USERPROFILE\Desktop\*.apk -Destination C:\inetpub\wwwroot\repos\api\Cloud.Gateway.Api\api\Atualizacao\etiquetas\ -Force
+            copy-item $env:USERPROFILE\Desktop\*.apk -Destination C:\inetpub\wwwroot\repos\api\Cloud.Gateway.Api\api\Atualizacao\ -Force
+			Remove-Item -Path $env:USERPROFILE\Desktop\*.apk -Force
+        }		
+    }
+    if($($Client -eq "centerlar") -and $($env:PDA_ENVIRONMENT -eq "prod")){
+        if($Project -eq "Admin" -or $Project -eq "admin" -or $Project -eq "ADMIN"){
+            Stop-IISSite -Name Centerlar.Admin.Front
+            Expand-Archive -Path $env:USERPROFILE\Desktop\admin.zip -DestinationPath C:\inetpub\wwwroot\repos\Centerlar\Cloud.Admin.Front -Force
+            Remove-Item -Path $env:USERPROFILE\Desktop\admin.zip -Force
+            Start-IISSite -Name Centerlar.Admin.Front
+        }
+		if($Project -eq "Etiquetas" -or $Project -eq "Label" -or $Project -eq "etq"){
+            Expand-Archive -Path $env:USERPROFILE\Desktop\etiquetas.zip -DestinationPath C:\inetpub\wwwroot\repos\Centerlar\Cloud.Gateway.Api\api\Atualizacao\etiquetas\ -Force    
+			Remove-Item -Path $env:USERPROFILE\Desktop\etiquetas.zip -Force
+        }
+		if($Project -eq "mobile" -or $Project -eq "apk"){
+            copy-item $env:USERPROFILE\Desktop\*.apk -Destination C:\inetpub\wwwroot\repos\api\Cloud.Gateway.Api\api\Atualizacao\ -Force
+			Remove-Item -Path $env:USERPROFILE\Desktop\*.apk -Force
+        }		
+    }
+    if($($Client -eq "centerlar") -and $($env:PDA_ENVIRONMENT -eq "homol")){
+        if($Project -eq "Admin" -or $Project -eq "admin" -or $Project -eq "ADMIN"){
+            Stop-IISSite -Name Centerlar.Admin.Front
+            Expand-Archive -Path $env:USERPROFILE\Desktop\admin.zip -DestinationPath C:\inetpub\wwwroot\repos\Centerlar\Cloud.Admin.Front -Force
+            Remove-Item -Path $env:USERPROFILE\Desktop\admin.zip -Force
+            Start-IISSite -Name Centerlar.Admin.Front
+        }
+		if($Project -eq "Etiquetas" -or $Project -eq "Label" -or $Project -eq "etq"){
+            Expand-Archive -Path $env:USERPROFILE\Desktop\etiquetas.zip -DestinationPath C:\inetpub\wwwroot\repos\Centerlar\Cloud.Gateway.Api\api\Atualizacao\etiquetas\ -Force    
+			Remove-Item -Path $env:USERPROFILE\Desktop\etiquetas.zip -Force
+        }
+		if($Project -eq "mobile" -or $Project -eq "apk"){
+            copy-item $env:USERPROFILE\Desktop\*.apk -Destination C:\inetpub\wwwroot\Centerlar\Cloud.Gateway.Api\api\Atualizacao\ -Force
 			Remove-Item -Path $env:USERPROFILE\Desktop\*.apk -Force
         }		
     }
